Inline one method on Runtime::ForProgrammingLanguages
