+++ lib/cucumber/runtime/for_programming_languages.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 require 'forwardable'
 require 'cucumber/core/ast/doc_string'
-require 'cucumber/core/ast/data_table'
 
 module Cucumber
   class Runtime
@@ -29,32 +28,6 @@ module Cucumber
         :invoke_dynamic_steps,
         :invoke_dynamic_step,
         :load_programming_language
-
-      # Returns a Cucumber::MultilineArgument::DataTable for +text_or_table+, which can either
-      # be a String:
-      #
-      #   table(%{
-      #     | account | description | amount |
-      #     | INT-100 | Taxi        | 114    |
-      #     | CUC-101 | Peeler      | 22     |
-      #   })
-      #
-      # or a 2D Array:
-      #
-      #   table([
-      #     %w{ account description amount },
-      #     %w{ INT-100 Taxi        114    },
-      #     %w{ CUC-101 Peeler      22     }
-      #   ])
-      #
-      def table(text_or_table, file=nil, line=0)
-        if !file
-          location = Core::Ast::Location.of_caller
-        else
-          location = Core::Ast::Location.new(file, line)
-        end
-        MultilineArgument::DataTable.from(text_or_table, location)
-      end
 
       # Returns a Cucumber::MultilineArgument::DocString for +body+.
       #
